Week3 : BaseballElimination: Remove edges connected to nodes involving games associated with target team.
Also, if a team is trivially eliminated, then the
certificate of elimination for that team is
contains just these teams and further investigation
as to the non-trivially eliminating teams is not
necessary. In fact, the non-trivial COE constitutes
of no specific tyeam in particular, and they must
be considered as a group as any one (at least) of them
will eliminate the target team but it is not known which.
Whereas with the trivial COE, each team can be considered
in isolation since they have already eliminated the
target team.